profiler-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-generativeai
python-dotenv
google-genai